requirements.txt
Flask==3.1.1
gunicorn==23.0.0
Werkzeug==3.1.3
pdf2image==1.17.0
python-pptx==1.0.2
PyPDF2==3.0.1
Pillow==11.3.0
email-validator==2.2.0